@@ -6,11 +6,11 @@ plugins {
 
 
 android {
-    compileSdk 28
+    compileSdk 31
 
     defaultConfig {
         minSdk 19
-        targetSdk 28
+        targetSdk 31
 
         consumerProguardFiles "consumer-rules.pro"
     }
@@ -31,9 +31,9 @@ android {
 }
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
-    implementation 'com.google.code.gson:gson:2.8.5'
-    implementation 'com.github.bumptech.glide:glide:4.8.0'
-    implementation 'com.squareup.okhttp3:okhttp:3.12.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation 'com.google.code.gson:gson:2.8.6'
+    implementation 'com.github.bumptech.glide:glide:4.9.0'
+    implementation 'com.squareup.okhttp3:okhttp:3.12.1'
 }
