@@ -7,12 +7,12 @@ plugins {
 }
 
 android {
-    compileSdk 28
+    compileSdk 31
 
     defaultConfig {
         applicationId "com.ct7ct7ct7.androidvimeoplayersample"
         minSdk 19
-        targetSdk 28
+        targetSdk 31
 
         consumerProguardFiles "consumer-rules.pro"
     }
@@ -40,7 +40,9 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(':library')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.android.support:design:28.0.0'
+    implementation 'androidx.appcompat:appcompat:1.4.1'
+    implementation 'androidx.drawerlayout:drawerlayout:1.1.1'
+    implementation 'com.google.android.material:material:1.5.0'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+
 }
