@@ -1,7 +1,7 @@
 package com.ct7ct7ct7.androidvimeoplayersample.examples
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.ct7ct7ct7.androidvimeoplayersample.R
 import com.ct7ct7ct7.androidvimeoplayersample.databinding.ActivityOriginalControlsBinding
 
