@@ -1,8 +1,8 @@
 package com.ct7ct7ct7.androidvimeoplayersample.examples
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
-import android.support.v7.widget.LinearLayoutManager
+import androidx.appcompat.app.AppCompatActivity
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.ct7ct7ct7.androidvimeoplayersample.R
 import com.ct7ct7ct7.androidvimeoplayersample.databinding.ActivityRecyclerviewBinding
 import java.util.ArrayList
