@@ -1,9 +1,6 @@
 package com.ct7ct7ct7.androidvimeoplayer.view.menu;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,6 +10,10 @@ import android.widget.TextView;
 import com.ct7ct7ct7.androidvimeoplayer.R;
 
 import java.util.List;
+
+import androidx.annotation.NonNull;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
 
 
 class MenuAdapter extends RecyclerView.Adapter<MenuAdapter.ViewHolder> {
