@@ -1,9 +1,10 @@
 package com.ct7ct7ct7.androidvimeoplayer.view.menu;
 
-import android.support.annotation.DrawableRes;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.View;
+
+import androidx.annotation.DrawableRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 
 public class ViemoMenuItem {
