@@ -1,9 +1,9 @@
 package com.ct7ct7ct7.androidvimeoplayersample.examples
 
-import android.arch.lifecycle.Lifecycle
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.lifecycle.Lifecycle
+import androidx.recyclerview.widget.RecyclerView
 import com.ct7ct7ct7.androidvimeoplayersample.databinding.ItemRecyclerviewBinding
 import java.util.ArrayList
 
@@ -11,7 +11,8 @@ class TestRecyclerViewAdapter(val lifecycle: Lifecycle, val items: ArrayList<Int
     RecyclerView.Adapter<TestRecyclerViewAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
-        val binding = ItemRecyclerviewBinding.inflate(LayoutInflater.from(parent.context), parent, false)
+        val binding =
+            ItemRecyclerviewBinding.inflate(LayoutInflater.from(parent.context), parent, false)
         lifecycle.addObserver(binding.vimeoPlayer)
         return ViewHolder(binding)
     }
